Remove build directory from repo
