+++ .gitignore
@@ -69,4 +69,5 @@ yarn.lock
 kalah-angular/kalah-angular.iml
 kalah-gameplay/kalah-gameplay.iml
 kalah-rest/kalah-rest.iml
+kalah-dao/kalah-dao.iml
 kalah.iml
