+++ kalah-dao/src/main/java/com/ttstudios/kalah/persistence/Application.java
@@ -1,0 +1,50 @@
+package com.ttstudios.kalah.persistence;
+
+import com.ttstudios.kalah.persistence.model.User;
+import com.ttstudios.kalah.persistence.repo.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application implements CommandLineRunner {
+
+    @Autowired
+    private UserRepository repository;
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+    @Override
+    public void run(String... args) throws Exception {
+
+        repository.deleteAll();
+
+        // save a couple of customers
+        repository.save(new User("Alice", "Smith", "AliceSmith"));
+        repository.save(new User("Bob", "Smith", "BobSmith"));
+
+        // fetch all customers
+        System.out.println("Customers found with findAll():");
+        System.out.println("-------------------------------");
+        for (User customer : repository.findAll()) {
+            System.out.println(customer);
+        }
+        System.out.println();
+
+        // fetch an individual customer
+        System.out.println("Customer found with findByFirstName('Alice'):");
+        System.out.println("--------------------------------");
+        System.out.println(repository.findByUserName("Alice"));
+
+        System.out.println("Customers found with findByLastName('Smith'):");
+        System.out.println("--------------------------------");
+        for (User customer : repository.findByUserName("Smith")) {
+            System.out.println(customer);
+        }
+
+    }
+
+}
