+++ kalah-dao/src/main/java/com/ttstudios/kalah/persistence/config/MongoConfig.java
@@ -1,0 +1,30 @@
+package com.ttstudios.kalah.persistence.config;
+
+import com.mongodb.Mongo;
+import com.mongodb.MongoClient;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.config.AbstractMongoConfiguration;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+/**
+ * Created by Timothy Tseng on 10-6-2017.
+ */
+@Configuration
+@EnableMongoRepositories(basePackages = "com.ttstudios.kalah.persistence.repo")
+public class MongoConfig extends AbstractMongoConfiguration {
+
+    @Override
+    protected String getDatabaseName() {
+        return "test";
+    }
+
+    @Override
+    public Mongo mongo() throws Exception {
+        return new MongoClient("127.0.0.1", 27017);
+    }
+
+    @Override
+    protected String getMappingBasePackage() {
+        return "com.ttstudios.kalah.persistence";
+    }
+}
