+++ kalah-dao/src/main/java/com/ttstudios/kalah/persistence/model/Book.java
@@ -1,0 +1,89 @@
+package com.ttstudios.kalah.persistence.model;
+
+import com.querydsl.core.annotations.QueryEntity;
+import org.bson.types.ObjectId;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import javax.validation.constraints.NotNull;
+
+@QueryEntity
+@Document
+public class Book {
+
+    @Id
+    private String id;
+
+    private String title;
+
+    @NotNull
+    private String author;
+
+    public Book() {
+        super();
+    }
+
+    public Book(String title, String author) {
+        super();
+        this.title = title;
+        this.author = author;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId( String id ) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+    @Override
+    public boolean equals( Object o ) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        Book book = (Book) o;
+
+        if (id != null ? !id.equals( book.id ) : book.id != null) {
+            return false;
+        }
+        if (title != null ? !title.equals( book.title ) : book.title != null) {
+            return false;
+        }
+        return author != null ? author.equals( book.author ) : book.author == null;
+    }
+
+    @Override
+    public int hashCode() {
+        int result = id != null ? id.hashCode() : 0;
+        result = 31 * result + (title != null ? title.hashCode() : 0);
+        result = 31 * result + (author != null ? author.hashCode() : 0);
+        return result;
+    }
+
+    @Override
+    public String toString() {
+        return "Book [id=" + id + ", title=" + title + ", author=" + author + "]";
+    }
+
+}
