+++ kalah-dao/src/main/java/com/ttstudios/kalah/persistence/model/State.java
@@ -1,0 +1,9 @@
+package com.ttstudios.kalah.persistence.model;
+
+import java.io.Serializable;
+
+public class State implements Serializable {
+
+    private static final long serialVersionUID = 6581287231361974540L;
+
+}
