+++ kalah-rest/pom.xml
@@ -20,7 +20,13 @@
     <dependencies>
 
         <dependency>
-            <groupId>com.ttstudios.kalah</groupId>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>kalah-dao</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>kalah-gameplay</artifactId>
             <version>${project.version}</version>
         </dependency>
@@ -33,10 +39,10 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-thymeleaf</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-jpa</artifactId>
-        </dependency>
+        <!--<dependency>-->
+            <!--<groupId>org.springframework.boot</groupId>-->
+            <!--<artifactId>spring-boot-starter-data-jpa</artifactId>-->
+        <!--</dependency>-->
 
         <!-- Websockets -->
         <dependency>
@@ -60,6 +66,13 @@
             <artifactId>jackson-databind</artifactId>
             <version>2.7.3</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.hibernate.validator</groupId>
+            <artifactId>hibernate-validator</artifactId>
+            <version>6.0.1.Final</version>
+        </dependency>
+
         <!-- DB -->
         <dependency>
             <groupId>com.h2database</groupId>
