--- kalah-rest/src/main/java/com/ttstudios/kalah/persistence/model/Message.java
@@ -1,32 +1,0 @@
-package com.ttstudios.kalah.persistence.model;
-
-public class Message {
-
-    private String message;
-    private int id;
-
-    public Message() {
-
-    }
-
-    public Message(int id, String message) {
-        this.id = id;
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-}
