--- kalah-rest/src/main/java/com/ttstudios/kalah/persistence/model/OutputMessage.java
@@ -1,21 +1,0 @@
-package com.ttstudios.kalah.persistence.model;
-
-import java.util.Date;
-
-public class OutputMessage extends Message {
-
-    private Date time;
-
-    public OutputMessage(Message original, Date time) {
-        super(original.getId(), original.getMessage());
-        this.time = time;
-    }
-
-    public Date getTime() {
-        return time;
-    }
-
-    public void setTime(Date time) {
-        this.time = time;
-    }
-}
