--- kalah-rest/src/main/java/com/ttstudios/kalah/persistence/model/State.java
@@ -1,4 +1,0 @@
-package com.ttstudios.kalah.persistence.model;
-
-public class State {
-}
