--- kalah-rest/src/main/java/com/ttstudios/kalah/persistence/repo/KalahRepository.java
@@ -1,10 +1,0 @@
-package com.ttstudios.kalah.persistence.repo;
-
-import com.ttstudios.kalah.persistence.model.KalahGame;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface KalahRepository extends CrudRepository<KalahGame, Long> {
-    List<KalahGame> findByTitle(String title);
-}
